fix(titus): everything was in the wrong package, breaking package scan autowiring
