+++ keel-titus-plugin/src/main/kotlin/com/netflix/spinnaker/keel/titus/exceptions/NoDockerImageFound.kt
@@ -15,11 +15,13 @@
  * limitations under the License.
  *
  */
-package com.netflix.spinnaker.titus.exceptions
+package com.netflix.spinnaker.keel.titus.exceptions
 
-import com.netflix.spinnaker.keel.docker.ContainerProvider
-import com.netflix.spinnaker.kork.exceptions.SystemException
+import com.netflix.spinnaker.keel.core.ResourceCurrentlyUnresolvable
+import com.netflix.spinnaker.kork.exceptions.IntegrationException
 
-class ErrorResolvingContainerException(
-  val container: ContainerProvider
-) : SystemException("There was an error resolving the correct docker container (current container: $container)")
+class NoDigestFound(repository: String, tag: String) :
+  ResourceCurrentlyUnresolvable("No digest found for docker image $repository:$tag in any registry")
+
+class RegistryNotFound(titusAccount: String) :
+  IntegrationException("Unable to find docker registry for titus account $titusAccount")
