+++ keel-titus-plugin/src/test/kotlin/com/netflix/spinnaker/keel/titus/resource/TestUtils.kt
@@ -12,8 +12,8 @@ import com.netflix.spinnaker.keel.clouddriver.model.ServiceJobProcesses
 import com.netflix.spinnaker.keel.clouddriver.model.TitusActiveServerGroup
 import com.netflix.spinnaker.keel.clouddriver.model.TitusActiveServerGroupImage
 import com.netflix.spinnaker.keel.core.parseMoniker
-import com.netflix.spinnaker.titus.CLOUD_PROVIDER
-import com.netflix.spinnaker.titus.moniker
+import com.netflix.spinnaker.keel.titus.CLOUD_PROVIDER
+import com.netflix.spinnaker.keel.titus.moniker
 import org.apache.commons.lang3.RandomStringUtils
 import com.netflix.spinnaker.keel.clouddriver.model.InstanceCounts as ClouddriverInstanceCounts
 import com.netflix.spinnaker.keel.clouddriver.model.TitusServerGroup as ClouddriverTitusServerGroup
